+++ package.json
@@ -21,6 +21,7 @@
     "@svgr/webpack": "^5.5.0",
     "clsx": "^1.1.1",
     "docusaurus-lunr-search": "^2.1.15",
+    "docusaurus-node-polyfills": "^1.0.0",
     "file-loader": "^6.2.0",
     "prism-react-renderer": "^1.2.1",
     "react": "^17.0.1",
