+++ src/theme/DocSidebarItem/Link/styles.module.css
@@ -5,6 +5,11 @@
  * LICENSE file in the root directory of this source tree.
  */
 
+.menuExternalLink {
+  align-items: center;
+}
+
+
 .menuLink {
   position: relative;
 }
