+++ src/theme/SearchBar/algolia.css
