+++ src/theme/SearchBar/lib/DocSearch.js
