+++ src/theme/SearchBar/lib/lunar-search.js
