+++ src/theme/SearchBar/lib/templates.js
