+++ src/theme/SearchBar/lib/utils.js
