+++ src/theme/SearchBar/styles.css
