--- src/theme_old/ApiDoc/index.ts
@@ -1,3 +1,0 @@
-import ApiDoc from './ApiDoc';
-
-export default ApiDoc;
