--- src/theme_old/ApiSchema/index.ts
@@ -1,3 +1,0 @@
-import ApiSchema from './ApiSchema';
-
-export default ApiSchema;
