--- src/theme_old/ApiSchema/styles.css
@@ -1,6 +1,0 @@
-.redocusaurus-schema.hide-example > div > div > div:nth-child(2) {
-  display: none;
-}
-.redocusaurus-schema.hide-example > div > div > div:nth-child(1) {
-  width: 100%;
-}
