--- src/theme_old/DocItem/styles.module.css
@@ -1,58 +1,0 @@
-/**
- * Copyright (c) Facebook, Inc. and its affiliates.
- *
- * This source code is licensed under the MIT license found in the
- * LICENSE file in the root directory of this source tree.
- */
-
-.docItemContainer article > *:first-child,
-.docItemContainer header + * {
-  margin-top: 0;
-}
-
-.docAuthors {
-  border-top: 1px solid var(--ifm-color-dark);
-  border-bottom: 1px solid var(--ifm-color-dark);
-  padding-top: 1rem;
-  padding-bottom: 1rem;
-  margin-top: 1rem;
-  margin-bottom: 1rem;
-  display: flex;
-  flex-wrap: wrap;
-  align-items: center;
-}
-
-.docAuthors > * {
-  margin-right: 0.5rem;
-}
-
-.authorByline {
-  margin-top: 6px;
-  display: flex;
-  align-items: center;
-}
-
-.authorLabel {
-  padding-top: 5px;
-  line-height: 0.5;
-  font-weight: 600;
-}
-
-.authorProfileImage {
-  display: inline-block;
-  border-radius: 50%;
-  width: 40px;
-  height: 40px;
-  margin-right: 8px;
-}
-
-@media (min-width: 997px) {
-  .docItemCol {
-    max-width: 75% !important;
-  }
-
-  /* Prevent hydration FOUC, as the mobile TOC needs to be server-rendered */
-  .tocMobile {
-    display: none;
-  }
-}
